fix(rtk): make MeshKit error test runnable and wire it into CI
The #1087 regression test imported ./api.ts, which value-imports
createApi/fetchBaseQuery from @reduxjs/toolkit/query/react; that entry
transitively requires react-redux (a peer dep absent in this repo's
test/CI env), so `npm test` died at module resolution (ERR_MODULE_NOT_FOUND)
and the guard never executed a single assertion.

Extract the pure MeshKit error surface (MeshkitError,
MeshkitFetchBaseQueryError, withMeshkitErrorTransform) into
typescript/rtk/meshkitError.ts, whose only RTK dependency is an
`import type` (erased at runtime). api.ts re-exports it so
@meshery/schemas/api consumers are unchanged (verified: built dist/api.d.ts
still exports all three). The test targets meshkitError.ts and runs under
Node's native TS type-stripping with no react-redux -- 4/4 pass.

Wire the guard into CI (addresses Copilot review): add a test-ts make
target and run it in generate-artifacts-from-schemas.yml, pinning
node-version 22 (>= 22.6 for node --test type-stripping). Exclude
*.test.ts from tsconfig -- test files need explicit .ts import extensions
for Node ESM, which tsc's node moduleResolution rejects (TS5097).
