+++ .github/workflows/generate-artifacts-from-schemas.yml
@@ -29,6 +29,9 @@ jobs:
 
       - name: Install build dependencies
         uses: actions/setup-node@v7
+        with:
+          # >= 22.6 for `node --test` native TypeScript type-stripping (make test-ts).
+          node-version: '22'
       - run: |
           npm i
           go version 
@@ -45,6 +48,11 @@ jobs:
         shell: bash
         run: |
           make site-data-test
+
+      - name: TypeScript unit tests
+        shell: bash
+        run: |
+          make test-ts
 
       - name: Site data generation
         shell: bash
